move custom_toast to utils (#1195)
